removed unused components
