+++ src/components/location/LocationSearchBar.js
@@ -3,7 +3,31 @@ import PlacesAutocomplete, {
   geocodeByAddress,
   getLatLng
 } from 'react-places-autocomplete';
-import { classnames } from './utils';
+
+const isObject = val => {
+  return typeof val === 'object' && val !== null;
+};
+
+const classnames = (...args) => {
+  const classes = [];
+  args.forEach(arg => {
+    if (typeof arg === 'string') {
+      classes.push(arg);
+    } else if (isObject(arg)) {
+      Object.keys(arg).forEach(key => {
+        if (arg[key]) {
+          classes.push(key);
+        }
+      });
+    } else {
+      throw new Error(
+        '`classnames` only accepts string or object as arguments'
+      );
+    }
+  });
+
+  return classes.join(' ');
+};
 
 class LocationSearchBar extends Component {
   constructor(props) {
