+++ src/components/location/maps/StaticMap.js
